+++ packages/css-tokenizer/src/index.ts
@@ -30,7 +30,7 @@
  * @packageDocumentation
  */
 export type { CodePointReader } from './interfaces/code-point-reader';
-export { ParseError, ParseErrorMessage } from './interfaces/error';
+export { ParseError, ParseErrorWithToken, ParseErrorMessage } from './interfaces/error';
 export { Reader } from './reader';
 export { cloneTokens } from './util/clone-tokens';
 export { stringify } from './stringify';