+++ packages/css-tokenizer/src/tokenizer.ts
@@ -15,7 +15,7 @@ import { Reader } from './reader';
 import { consumeStringToken } from './consume/string-token';
 import { consumeIdentLikeToken } from './consume/ident-like-token';
 import { checkIfTwoCodePointsAreAValidEscape } from './checks/two-code-points-are-valid-escape';
-import { ParseError, ParseErrorMessage } from './interfaces/error';
+import { ParseError, ParseErrorMessage, ParseErrorWithToken } from './interfaces/error';
 import { checkIfThreeCodePointsWouldStartAUnicodeRange } from './checks/three-code-points-would-start-unicode-range';
 import { consumeUnicodeRangeToken } from './consume/unicode-range-token';
 
@@ -229,7 +229,7 @@ export function tokenizer(
 					value: '\\',
 				}];
 
-				ctx.onParseError(new ParseError(
+				ctx.onParseError(new ParseErrorWithToken(
 					ParseErrorMessage.InvalidEscapeSequenceAfterBackslash,
 					reader.representationStart,
 					reader.representationEnd,